changed module.export -> exports.signup on controllers/AgentSignup.js rename Controllers/AgentSignupController.js -> controllers/AgentSignup.js rename Routes/AgentRoute.js -> routes/AgentSignup.js change module.export -> exports.signup on controllers/AgentSignup.js on router.post in routes/AgentSignup.js, changed AgentSignup -> AgentSignup.signup
